add routes PATCH /v3/me/data + PATCH /v3/users/:id/data for setting generic user data
