+++ src/routes/me.js
@@ -63,30 +63,11 @@ module.exports = {
             handler: updateMe
         });
 
-        server.route({
-            method: 'PATCH',
-            path: '/settings',
-            options: {
-                tags: ['api'],
-                description: 'Update your account settings',
-                notes: 'Use this endpoint to change your active team.',
-                validate: {
-                    payload: {
-                        activeTeam: Joi.string()
-                            .allow(null)
-                            .example('teamxyz')
-                            .description('Your active team')
-                    }
-                },
-                response: createResponseConfig({
-                    schema: Joi.object({
-                        activeTeam: Joi.string(),
-                        updatedAt: Joi.date()
-                    }).unknown()
-                })
-            },
-            handler: updateMySettings
-        });
+        // GET /v3/me/:id/settings
+        require('./me/settings')(server, options);
+
+        // GET /v3/me/:id/data
+        require('./me/data')(server, options);
 
         server.route({
             method: 'DELETE',
@@ -140,17 +121,6 @@ async function updateMe(request, h) {
     const res = await request.server.inject({
         method: 'PATCH',
         url: `/v3/users/${request.auth.artifacts.id}`,
-        auth: request.auth,
-        payload: request.payload
-    });
-
-    return h.response(res.result).code(res.statusCode);
-}
-
-async function updateMySettings(request, h) {
-    const res = await request.server.inject({
-        method: 'PATCH',
-        url: `/v3/users/${request.auth.artifacts.id}/settings`,
         auth: request.auth,
         payload: request.payload
     });
